requirements.txt
flask
flask-cors
flask-sqlalchemy
yt-dlp
google-generativeai
werkzeug
gunicorn

